% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeScreener-package.R
\docType{package}
\name{CopeScreener-package}
\alias{CopeScreener}
\alias{CopeScreener-package}
\title{CopeScreener: COPE Screener API Package}
\description{
Pulls and cleans cope screener responses
}
\author{
\strong{Maintainer}: Amy Hume \email{amyhume66@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
