% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cope_screener_functions.r
\name{get_cope_screener}
\alias{get_cope_screener}
\title{COPE SCREENER PACKAGE}
\usage{
get_cope_screener(
  token = token,
  form = "cope_screener_survey",
  raw_v_label = "raw"
)
}
\arguments{
\item{token}{The API token for the project}

\item{form}{The name of the REDCap form to retrieve data from}

\item{raw_v_label}{The label for raw data fields}
}
\value{
A data frame with the cope_screener_survey form
}
\description{
This function retrieves data from the COPE Screener REDCap project using the API.
}
