% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cope_screener_functions.r
\name{get_cope_screener_clean}
\alias{get_cope_screener_clean}
\title{This function retrieves cleaned data from the COPE Screener REDCap project using the API.}
\usage{
get_cope_screener_clean(token, min_date = "2020-01-01")
}
\arguments{
\item{token}{The API token for the project}

\item{min_date}{The minimum date you want to pull screener responses for. Default is all}
}
\value{
A data frame with the cleaned screener responses for your specified date range
}
\description{
This function retrieves cleaned data from the COPE Screener REDCap project using the API.
}
